+++ src/api/model/streams/mod.rs
@@ -5,3 +5,4 @@ pub(in crate::api) mod shared_stream_manager;
 pub(in crate::api) mod active_client_stream;
 mod buffered_stream;
 mod client_stream;
+mod freeze_frame_stream;