+++ src/api/model/streams/provider_stream.rs
@@ -1,19 +1,20 @@
 use crate::api::api_utils::{get_headers_from_request, HeaderFilter};
+use crate::api::model::model_utils::get_response_headers;
+use crate::api::model::stream_error::StreamError;
+use crate::api::model::streams::freeze_frame_stream::FreezeFrameStream;
 use crate::api::model::streams::provider_stream_factory::{create_provider_stream, BufferStreamOptions};
-use crate::model::config::ConfigInput;
+use crate::model::config::{Config, ConfigInput};
+use crate::model::playlist::PlaylistItemType;
+use crate::utils::debug_if_enabled;
 use crate::utils::network::request::{get_request_headers, sanitize_sensitive_info};
-use actix_web::{HttpRequest};
+use actix_web::HttpRequest;
 use bytes::Bytes;
 use futures::stream::BoxStream;
+use futures::TryStreamExt;
 use log::error;
 use reqwest::StatusCode;
 use std::sync::Arc;
-use futures::TryStreamExt;
 use url::Url;
-use crate::api::model::model_utils::get_response_headers;
-use crate::api::model::stream_error::StreamError;
-use crate::model::playlist::PlaylistItemType;
-use crate::utils::debug_if_enabled;
 
 type ProviderStreamResponse = (Option<BoxStream<'static, Result<Bytes, StreamError>>>, Option<(Vec<(String, String)>, StatusCode)>);
 
@@ -27,7 +28,8 @@ pub fn get_header_filter_for_item_type(item_type: PlaylistItemType) -> HeaderFil
     }
 }
 
-pub async fn get_provider_pipe_stream(http_client: &Arc<reqwest::Client>,
+pub async fn get_provider_pipe_stream(cfg: &Config,
+                                      http_client: &Arc<reqwest::Client>,
                                       stream_url: &Url,
                                       req: &HttpRequest,
                                       input: Option<&ConfigInput>,
@@ -46,7 +48,9 @@ pub async fn get_provider_pipe_stream(http_client: &Arc<reqwest::Client>,
             let response_headers = get_response_headers(&mut response);
             let status = response.status();
             if status.is_success() {
-                (Some(Box::pin(response.bytes_stream().map_err(|err|StreamError::reqwest(&err)))), Some((response_headers, status)))
+                (Some(Box::pin(response.bytes_stream().map_err(|err| StreamError::reqwest(&err)))), Some((response_headers, status)))
+            } else if let Some(freeze_frame) = cfg.t_channel_unavailable_file.as_ref() {
+                (Some(Box::pin(FreezeFrameStream::new(status.as_u16(), Arc::clone(freeze_frame)))), Some((response_headers, status)))
             } else {
                 (None, Some((response_headers, status)))
             }
@@ -59,12 +63,13 @@ pub async fn get_provider_pipe_stream(http_client: &Arc<reqwest::Client>,
     }
 }
 
-pub async fn get_provider_reconnect_buffered_stream(http_client: &Arc<reqwest::Client>,
+pub async fn get_provider_reconnect_buffered_stream(cfg: &Config,
+                                                    http_client: &Arc<reqwest::Client>,
                                                     stream_url: &Url,
                                                     req: &HttpRequest,
                                                     input: Option<&ConfigInput>,
                                                     options: BufferStreamOptions) -> ProviderStreamResponse {
-    match create_provider_stream(Arc::clone(http_client), stream_url, req, input, options).await {
+    match create_provider_stream(cfg, Arc::clone(http_client), stream_url, req, input, options).await {
         None => (None, None),
         Some((stream, info)) => {
             (Some(stream), info)